+++ src/reasoner.rs
